Move partition logic into ugrid base class Add DataArray method for convenience
